chore!: rename binary from ecs-preview to copilot (#950)
