+++ NOTICE
@@ -1,2 +1,2 @@
-Amazon ECS ecs-preview
-Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved. 
+Amazon ECS Copilot
+Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
