+++ internal/pkg/cli/cli.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package cli contains the ecs-preview subcommands.
+// Package cli contains the copilot subcommands.
 package cli
 
 import (
