+++ internal/pkg/cli/completion.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/aws/amazon-ecs-cli-v2/cmd/ecs-preview/template"
+	"github.com/aws/amazon-ecs-cli-v2/cmd/copilot/template"
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/cli/group"
 )
 
@@ -56,18 +56,18 @@ func BuildCompletionCmd(rootCmd *cobra.Command) *cobra.Command {
 The code must be evaluated to provide interactive completion of commands.`,
 		Example: `
   Install zsh completion
-  /code $ source <(ecs-preview completion zsh)
-  /code $ ecs-preview completion zsh > "${fpath[1]}/_ecs-preview" # to autoload on startup
+  /code $ source <(copilot completion zsh)
+  /code $ copilot completion zsh > "${fpath[1]}/_copilot" # to autoload on startup
 
   Install bash completion on macOS using homebrew
   /code $ brew install bash-completion   # if running 3.2
   /code $ brew install bash-completion@2 # if running Bash 4.1+
-  /code $ ecs-preview completion bash > /usr/local/etc/bash_completion.d
+  /code $ copilot completion bash > /usr/local/etc/bash_completion.d
 
   Install bash completion on linux
-  /code $ source <(ecs-preview completion bash)
-  /code $ ecs-preview completion bash > ecs-preview.sh
-  /code $ sudo mv ecs-preview.sh /etc/bash_completion.d/ecs-preview`,
+  /code $ source <(copilot completion bash)
+  /code $ copilot completion bash > copilot.sh
+  /code $ sudo mv copilot.sh /etc/bash_completion.d/copilot`,
 		Args: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 1 {
 				return errors.New("requires a single shell argument (bash or zsh)")
